refactor/#61: RefreshTokenRepository 추상화
