+++ src/main/java/com/tookscan/tookscan/security/application/service/ReissueJsonWebTokenService.java
@@ -1,7 +1,5 @@
 package com.tookscan.tookscan.security.application.service;
 
-import com.tookscan.tookscan.core.exception.error.ErrorCode;
-import com.tookscan.tookscan.core.exception.type.CommonException;
 import com.tookscan.tookscan.core.utility.JsonWebTokenUtil;
 import com.tookscan.tookscan.security.application.dto.response.DefaultJsonWebTokenDto;
 import com.tookscan.tookscan.security.application.usecase.ReissueJsonWebTokenUseCase;
@@ -9,7 +7,7 @@ import com.tookscan.tookscan.security.domain.mysql.Account;
 import com.tookscan.tookscan.security.domain.redis.RefreshToken;
 import com.tookscan.tookscan.security.domain.service.RefreshTokenService;
 import com.tookscan.tookscan.security.repository.AccountRepository;
-import com.tookscan.tookscan.security.repository.redis.RefreshTokenRepository;
+import com.tookscan.tookscan.security.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -33,8 +31,7 @@ public class ReissueJsonWebTokenService implements ReissueJsonWebTokenUseCase {
     public DefaultJsonWebTokenDto execute(String refreshTokenValue) {
 
         // refresh Token 검증. Redis에 있는 토큰인지 확인 -> accountId 추출
-        RefreshToken refreshToken = refreshTokenRepository.findByValue(refreshTokenValue)
-                .orElseThrow(() -> new CommonException(ErrorCode.INVALID_TOKEN_ERROR));
+        RefreshToken refreshToken = refreshTokenRepository.findByValueOrElseThrow(refreshTokenValue);
         UUID accountId = refreshToken.getAccountId();
 
         // Account 조회
