+++ src/main/java/com/tookscan/tookscan/security/repository/RefreshTokenRepository.java
@@ -1,0 +1,14 @@
+package com.tookscan.tookscan.security.repository;
+
+import com.tookscan.tookscan.security.domain.redis.RefreshToken;
+
+import java.util.UUID;
+
+public interface RefreshTokenRepository {
+
+    void save(RefreshToken refreshToken);
+
+    void deleteById(UUID value);
+
+    RefreshToken findByValueOrElseThrow(String value);
+}
